+++ src/cmd/replica/create.rs
@@ -5,10 +5,7 @@
 
 use crate::cmd::RESOURCE_PATH_HELP;
 use crate::io::reduct::{build_client, parse_url_and_token};
-use crate::parse::widely_used_args::{
-    make_each_n, make_each_s, make_entries_arg, make_exclude_arg, make_include_arg, make_when_arg,
-    parse_label_args,
-};
+use crate::parse::widely_used_args::{make_each_n, make_each_s, make_entries_arg, make_when_arg};
 use crate::parse::ResourcePathParser;
 use clap::{Arg, Command};
 use reduct_rs::Labels;
@@ -34,8 +31,6 @@ pub(super) fn create_replica_cmd() -> Command {
                 .value_parser(ResourcePathParser::new())
                 .required(true),
         )
-        .arg(make_include_arg())
-        .arg(make_exclude_arg())
         .arg(make_entries_arg())
         .arg(make_each_n())
         .arg(make_each_s())
@@ -59,8 +54,6 @@ pub(super) async fn create_replica(
         .map(|s| s.to_string())
         .collect::<Vec<String>>();
 
-    let include = parse_label_args(args.get_many::<String>("include"))?.unwrap_or_default();
-    let exclude = parse_label_args(args.get_many::<String>("exclude"))?.unwrap_or_default();
     let each_n = args.get_one::<u64>("each-n");
     let each_s = args.get_one::<f64>("each-s");
     let when = args.get_one::<String>("when");
@@ -74,8 +67,6 @@ pub(super) async fn create_replica(
         .dst_bucket(dest_bucket_name)
         .dst_host(dest_url.as_str())
         .dst_token(&token)
-        .include(Labels::from_iter(include))
-        .exclude(Labels::from_iter(exclude))
         .entries(entries_filter);
 
     if let Some(when) = when {
@@ -123,10 +114,6 @@ mod tests {
             format!("local/{}", test_replica).as_str(),
             &bucket,
             format!("local/{}", bucket2).as_str(),
-            "--include",
-            "key1=value2",
-            "--exclude",
-            "key2=value3",
             "--entries",
             "entry1",
             "entry2",
@@ -144,18 +131,6 @@ mod tests {
         assert_eq!(replica.settings.dst_bucket, bucket2);
         assert_eq!(replica.settings.dst_host, "http://localhost:8383/");
         assert_eq!(replica.settings.dst_token, "***");
-        assert_eq!(
-            replica.settings.include,
-            Labels::from_iter(vec![("key1".to_string(), "value2".to_string())])
-        );
-        assert_eq!(
-            replica.settings.exclude,
-            Labels::from_iter(vec![("key2".to_string(), "value3".to_string())])
-        );
-        assert_eq!(
-            replica.settings.entries,
-            vec!["entry1".to_string(), "entry2".to_string()]
-        );
         assert_eq!(replica.settings.each_n, Some(10));
         assert_eq!(replica.settings.each_s, Some(0.5));
         assert_eq!(