+++ src/parse/widely_used_args.rs
@@ -8,31 +8,11 @@ use clap::ArgAction::SetTrue;
 use clap::{value_parser, Arg};
 use reduct_rs::Labels;
 
-pub(crate) fn make_include_arg() -> Arg {
-    Arg::new("include")
-        .long("include")
-        .short('I')
-        .value_name("KEY=VALUE")
-        .help("Deprecated. Use --when instead.\nThese records which have this key-value pair will be requested.\nThe format is key=value. This option can be used multiple times to include multiple key-value pairs.")
-        .num_args(0..)
-        .required(false)
-}
-
-pub(crate) fn make_exclude_arg() -> Arg {
-    Arg::new("exclude")
-        .long("exclude")
-        .short('E')
-        .value_name("KEY=VALUE")
-        .help("Deprecated. Use --when instead.\nThese records which have this key-value pair will be excluded.\nThe format is key=value. This option can be used multiple times to exclude multiple key-value pairs.")
-        .num_args(0..)
-        .required(false)
-}
-
 pub(crate) fn make_each_n() -> Arg {
     Arg::new("each-n")
         .long("each-n")
         .short('N')
-        .help("Export every nth record.\nIf not specified, all records will be requested.")
+        .help("(DEPRECATED) Export every nth record.\nIf not specified, all records will be requested.")
         .value_name("NUMBER")
         .value_parser(value_parser!(u64))
         .required(false)
@@ -42,7 +22,7 @@ pub(crate) fn make_each_s() -> Arg {
     Arg::new("each-s")
         .long("each-s")
         .short('S')
-        .help("Export a record every n seconds.\nIf not specified, all records will be requested.")
+        .help("(DEPRECATED) Export a record every n seconds.\nIf not specified, all records will be requested.")
         .value_name("NUMBER")
         .value_parser(value_parser!(f64))
         .required(false)
@@ -97,18 +77,4 @@ pub(crate) fn parse_label(label: &str) -> anyhow::Result<(String, String)> {
             .ok_or(anyhow::anyhow!("Invalid label"))?
             .to_string(),
     ))
-}
-
-pub(crate) fn parse_label_args(args: Option<ValuesRef<String>>) -> anyhow::Result<Option<Labels>> {
-    let mut result: Option<Labels> = None;
-    if let Some(include_args) = args {
-        let mut labels = Labels::new();
-        for arg in include_args {
-            let (key, value) = parse_label(&arg)?;
-            let _ = labels.insert(key, value);
-        }
-
-        result = Some(labels);
-    }
-    Ok(result)
 }